Fixed buildings breaking after creating a few
For some reason the ENT.CreatedNetworkVars table wasn't being reset to 0
for each entity, so every time a building was spawned the slot numbers
would increment until they were above the limit of 32, breaking the
building.

Although the table is reset properly for the weapons, it's been changed
to the same thing the buildings now use, just in case.